+++ src/modules/0.0.21/aws_api_gateway/entity/api.ts
@@ -51,6 +51,5 @@ export class Api {
     nullable: false,
     default: () => 'default_aws_region()',
   })
-  @cloudId
   region: string;
 }
