+++ src/modules/0.0.21/aws_api_gateway/index.ts
@@ -2,10 +2,11 @@ import {
   ApiGatewayV2,
   Api as ApiAWS,
   CreateApiCommandInput,
+  CreateApiCommandOutput,
   UpdateApiCommandInput,
 } from '@aws-sdk/client-apigatewayv2';
 
-import { AWS, crudBuilder2, crudBuilderFormat } from '../../../services/aws_macros';
+import { AWS, crudBuilder2, crudBuilderFormat, paginateBuilder } from '../../../services/aws_macros';
 import { Context, Crud2, MapperBase, ModuleBase } from '../../interfaces';
 import { Api, Protocol } from './entity';
 
@@ -18,7 +19,8 @@ class ApiMapper extends MapperBase<Api> {
       Object.is(a.disableExecuteApiEndpoint, b.disableExecuteApiEndpoint) &&
       Object.is(a.protocolType, b.protocolType) &&
       Object.is(a.version, b.version) &&
-      Object.is(a.name, b.name);
+      Object.is(a.name, b.name) &&
+      Object.is(a.region, b.region);
     return res;
   };
 
@@ -66,33 +68,36 @@ class ApiMapper extends MapperBase<Api> {
       }
       return out;
     },
-    read: async (ctx: Context, id?: string) => {
+    read: async (ctx: Context, apiId?: string) => {
       const enabledRegions = (await ctx.getEnabledAwsRegions()) as string[];
-      if (!!id) {
-        const [apiId, region] = id.split('|');
-        if (enabledRegions.includes(region)) {
+      if (apiId) {
+        for (const region of enabledRegions) {
           const client = (await ctx.getAwsClient(region)) as AWS;
           const rawApi = await this.getApi(client.apiGatewayClient, apiId);
-          if (rawApi) return this.apiMapper(rawApi, region);
+          if (!rawApi) continue;
+          return this.apiMapper(rawApi, region);
         }
       } else {
-        const out: Api[] = [];
-        await Promise.all(enabledRegions.map(async region => {
+        const out = [];
+        for (const region of enabledRegions) {
           const client = (await ctx.getAwsClient(region)) as AWS;
           const rawApis = (await this.getApis(client.apiGatewayClient)) ?? [];
           for (const i of rawApis) {
             const outApi = this.apiMapper(i, region);
             if (outApi) out.push(outApi);
           }
-        }))
+        }
         return out;
       }
     },
+    updateOrReplace: (a: Api, b: Api) => (a.region === b.region ? 'update' : 'replace'),
     update: async (rs: Api[], ctx: Context) => {
       const out: Api[] = [];
       for (const r of rs) {
         const client = (await ctx.getAwsClient(r.region)) as AWS;
-        const cloudRecord = ctx?.memo?.cloud?.Api?.[this.entityId(r)];
+        const cloudRecord = ctx?.memo?.cloud?.Api?.[r.apiId ?? ''];
+        const isUpdate = Object.is(this.module.api.cloud.updateOrReplace(cloudRecord, r), 'update');
+        if (isUpdate) {
           // api id is generated by aws, we cannot modify it
           if (cloudRecord.protocolType !== r.protocolType) {
             // restore
@@ -121,7 +126,11 @@ class ApiMapper extends MapperBase<Api> {
               throw new Error('Error updating API');
             }
           }
-        
+        } else {
+          const newApi: Api | Api[] | undefined = await this.module.api.cloud.create(r, ctx);
+          await this.module.api.cloud.delete(cloudRecord, ctx);
+          if (newApi && !Array.isArray(newApi)) out.push(newApi);
+        }
       }
       return out;
     },
