+++ src/modules/0.0.22/aws_api_gateway/entity/api.ts
@@ -51,5 +51,6 @@ export class Api {
     nullable: false,
     default: () => 'default_aws_region()',
   })
+  @cloudId
   region: string;
 }
