+++ src/modules/0.0.22/aws_api_gateway/index.ts
@@ -2,11 +2,10 @@ import {
   ApiGatewayV2,
   Api as ApiAWS,
   CreateApiCommandInput,
-  CreateApiCommandOutput,
   UpdateApiCommandInput,
 } from '@aws-sdk/client-apigatewayv2';
 
-import { AWS, crudBuilder2, crudBuilderFormat, paginateBuilder } from '../../../services/aws_macros';
+import { AWS, crudBuilder2, crudBuilderFormat } from '../../../services/aws_macros';
 import { Context, Crud2, MapperBase, ModuleBase } from '../../interfaces';
 import { Api, Protocol } from './entity';
 
@@ -19,8 +18,7 @@ class ApiMapper extends MapperBase<Api> {
       Object.is(a.disableExecuteApiEndpoint, b.disableExecuteApiEndpoint) &&
       Object.is(a.protocolType, b.protocolType) &&
       Object.is(a.version, b.version) &&
-      Object.is(a.name, b.name) &&
-      Object.is(a.region, b.region);
+      Object.is(a.name, b.name);
     return res;
   };
 
@@ -68,68 +66,62 @@ class ApiMapper extends MapperBase<Api> {
       }
       return out;
     },
-    read: async (ctx: Context, apiId?: string) => {
+    read: async (ctx: Context, id?: string) => {
       const enabledRegions = (await ctx.getEnabledAwsRegions()) as string[];
-      if (apiId) {
-        for (const region of enabledRegions) {
+      if (!!id) {
+        const [apiId, region] = id.split('|');
+        if (enabledRegions.includes(region)) {
           const client = (await ctx.getAwsClient(region)) as AWS;
           const rawApi = await this.getApi(client.apiGatewayClient, apiId);
-          if (!rawApi) continue;
-          return this.apiMapper(rawApi, region);
+          if (rawApi) return this.apiMapper(rawApi, region);
         }
       } else {
-        const out = [];
-        for (const region of enabledRegions) {
-          const client = (await ctx.getAwsClient(region)) as AWS;
-          const rawApis = (await this.getApis(client.apiGatewayClient)) ?? [];
-          for (const i of rawApis) {
-            const outApi = this.apiMapper(i, region);
-            if (outApi) out.push(outApi);
-          }
-        }
+        const out: Api[] = [];
+        await Promise.all(
+          enabledRegions.map(async region => {
+            const client = (await ctx.getAwsClient(region)) as AWS;
+            const rawApis = (await this.getApis(client.apiGatewayClient)) ?? [];
+            for (const i of rawApis) {
+              const outApi = this.apiMapper(i, region);
+              if (outApi) out.push(outApi);
+            }
+          }),
+        );
         return out;
       }
     },
-    updateOrReplace: (a: Api, b: Api) => (a.region === b.region ? 'update' : 'replace'),
     update: async (rs: Api[], ctx: Context) => {
       const out: Api[] = [];
       for (const r of rs) {
         const client = (await ctx.getAwsClient(r.region)) as AWS;
-        const cloudRecord = ctx?.memo?.cloud?.Api?.[r.apiId ?? ''];
-        const isUpdate = Object.is(this.module.api.cloud.updateOrReplace(cloudRecord, r), 'update');
-        if (isUpdate) {
-          // api id is generated by aws, we cannot modify it
-          if (cloudRecord.protocolType !== r.protocolType) {
-            // restore
-            cloudRecord.id = r.id;
-            await this.module.api.db.update(cloudRecord, ctx);
-            out.push(cloudRecord);
-          } else {
-            const input: UpdateApiCommandInput = {
-              ApiId: r.apiId,
-              Name: r.name,
-              Description: r.description,
-              DisableExecuteApiEndpoint: r.disableExecuteApiEndpoint,
-              Version: r.version,
-            };
-            const res = await this.updateApi(client.apiGatewayClient, input);
-            if (res) {
-              const newApi = this.apiMapper(res, r.region);
-              if (newApi) {
-                newApi.name = r.name;
-                newApi.id = r.id;
-                // Save the record back into the database to get the new fields updated
-                await this.module.api.db.update(newApi, ctx);
-                out.push(newApi);
-              }
-            } else {
-              throw new Error('Error updating API');
-            }
-          }
+        const cloudRecord = ctx?.memo?.cloud?.Api?.[this.entityId(r)];
+        // api id is generated by aws, we cannot modify it
+        if (cloudRecord.protocolType !== r.protocolType) {
+          // restore
+          cloudRecord.id = r.id;
+          await this.module.api.db.update(cloudRecord, ctx);
+          out.push(cloudRecord);
         } else {
-          const newApi: Api | Api[] | undefined = await this.module.api.cloud.create(r, ctx);
-          await this.module.api.cloud.delete(cloudRecord, ctx);
-          if (newApi && !Array.isArray(newApi)) out.push(newApi);
+          const input: UpdateApiCommandInput = {
+            ApiId: r.apiId,
+            Name: r.name,
+            Description: r.description,
+            DisableExecuteApiEndpoint: r.disableExecuteApiEndpoint,
+            Version: r.version,
+          };
+          const res = await this.updateApi(client.apiGatewayClient, input);
+          if (res) {
+            const newApi = this.apiMapper(res, r.region);
+            if (newApi) {
+              newApi.name = r.name;
+              newApi.id = r.id;
+              // Save the record back into the database to get the new fields updated
+              await this.module.api.db.update(newApi, ctx);
+              out.push(newApi);
+            }
+          } else {
+            throw new Error('Error updating API');
+          }
         }
       }
       return out;
